+++ aspnetcore/security/key-vault-configuration/sample/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "KeyVaultName": "Key Vault Name",
+  "AzureADApplicationId": "Azure AD Application ID",
+  "AzureADPassword": "Azure AD Password/Client Secret"
+}
