--- aspnetcore/security/key-vault-configuration/samples/basic-sample/2.x/Program.cs
@@ -1,29 +1,0 @@
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-
-namespace KeyVaultConfigProviderSample
-{
-    public static class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateWebHostBuilder(args).Build().Run();
-        }
-
-        #region snippet1
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .ConfigureAppConfiguration((context, config) =>
-                {
-                    var builtConfig = config.Build();
-
-                    config.AddAzureKeyVault(
-                        $"https://{builtConfig["Vault"]}.vault.azure.net/",
-                        builtConfig["ClientId"],
-                        builtConfig["ClientSecret"]);
-                })
-                .UseStartup<Startup>();
-        #endregion
-    }
-}
