--- aspnetcore/security/key-vault-configuration/samples/key-name-prefix-sample/1.x/Program.cs
@@ -1,24 +1,0 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.Logging;
-
-namespace KeyVaultConfigProviderSample
-{
-    public static class Program
-    {
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .ConfigureLogging(factory =>
-                {
-                    factory.AddConsole(LogLevel.Debug);
-                })
-                .UseKestrel()
-                .UseIISIntegration()
-                .UseStartup<Startup>()
-                .Build();
-
-            host.Run();
-        }
-    }
-}
