--- aspnetcore/security/key-vault-configuration/samples/key-name-prefix-sample/1.x/appsettings.json
@@ -1,5 +1,0 @@
-{
-  "Vault": "Azure Key Vault name (for example: contosovault)",
-  "ClientId": "Azure Active Directory Application Id (for example: 627e911e-43cc-61d4-992e-12db9c81b413)",
-  "ClientSecret": "Azure Active Directory Key (for example: g58K3dtg59o1Pa+e59v2Tx829w6VxTB2yv9sv/101di=)"
-}
