--- aspnetcore/security/key-vault-configuration/samples/key-name-prefix-sample/2.x/PrefixKeyVaultConfigProviderSample.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFrameworks>netcoreapp2.1</TargetFrameworks>
-    <Version>5.0.0.0</Version>
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.AzureKeyVault" Version="2.1.0" />
-  </ItemGroup>
-
-</Project>
